SCB-910 refactor add spring-boot2-starter,spring-boot2-starter-standalone,spring-boot2-starter-servlet for support spring boot 2.x;add standalone deploy and servlet deploy in it
