+++ integration-tests/it-producer-deploy-springboot2-servelt/pom.xml
@@ -19,41 +19,34 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <groupId>org.apache.servicecomb</groupId>
-    <artifactId>java-chassis-dependencies-springboot2</artifactId>
-    <version>1.1.0-SNAPSHOT</version>
-    <relativePath>../java-chassis-dependencies-springboot2</relativePath>
-  </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>spring-boot2-starter-parent</artifactId>
-  <packaging>pom</packaging>
+  <groupId>org.apache.servicecomb.tests</groupId>
+  <artifactId>it-producer-deploy-springboot2-servelt</artifactId>
   <version>1.1.0-SNAPSHOT</version>
 
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <java.version>1.8</java.version>
-    <argLine>-Dfile.encoding=UTF-8</argLine>
-  </properties>
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>org.apache.servicecomb</groupId>
+        <artifactId>java-chassis-dependencies-springboot2</artifactId>
+        <version>1.1.0-SNAPSHOT</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
 
-  <modules>
-    <module>spring-boot2-starter</module>
-    <module>spring-boot2-starter-servlet</module>
-    <module>spring-boot2-starter-standalone</module>
-  </modules>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
+  <dependencies>
+    <dependency>
+      <groupId>org.apache.servicecomb.tests</groupId>
+      <artifactId>it-producer</artifactId>
+      <version>1.1.0-SNAPSHOT</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.servicecomb</groupId>
+      <artifactId>spring-boot2-starter-servlet</artifactId>
+    </dependency>
+  </dependencies>
 
 </project>
