+++ java-chassis-dependencies-springboot2/pom.xml
@@ -153,7 +153,17 @@
 
       <dependency>
         <groupId>org.apache.servicecomb</groupId>
+        <artifactId>spring-boot2-starter</artifactId>
+        <version>1.1.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.servicecomb</groupId>
         <artifactId>spring-boot2-starter-servlet</artifactId>
+        <version>1.1.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.servicecomb</groupId>
+        <artifactId>spring-boot2-starter-standalone</artifactId>
         <version>1.1.0-SNAPSHOT</version>
       </dependency>
