+++ spring-boot2-starter-parent/spring-boot2-starter-standalone/pom.xml
@@ -20,40 +20,34 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
+    <artifactId>spring-boot2-starter-parent</artifactId>
     <groupId>org.apache.servicecomb</groupId>
-    <artifactId>java-chassis-dependencies-springboot2</artifactId>
     <version>1.1.0-SNAPSHOT</version>
-    <relativePath>../java-chassis-dependencies-springboot2</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>spring-boot2-starter-parent</artifactId>
-  <packaging>pom</packaging>
-  <version>1.1.0-SNAPSHOT</version>
+  <artifactId>spring-boot2-starter-standalone</artifactId>
 
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <java.version>1.8</java.version>
-    <argLine>-Dfile.encoding=UTF-8</argLine>
-  </properties>
+  <dependencies>
+    <dependency>
+      <groupId>org.apache.servicecomb</groupId>
+      <artifactId>spring-boot2-starter</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.servicecomb</groupId>
+      <artifactId>transport-rest-vertx</artifactId>
+    </dependency>
 
-  <modules>
-    <module>spring-boot2-starter</module>
-    <module>spring-boot2-starter-servlet</module>
-    <module>spring-boot2-starter-standalone</module>
-  </modules>
+    <!--ConditionEvaluationReportLoggingListener need in autoconfigure-->
+    <dependency>
+      <groupId>org.hibernate</groupId>
+      <artifactId>hibernate-validator</artifactId>
+    </dependency>
 
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
+    <dependency>
+      <groupId>org.apache.servicecomb</groupId>
+      <artifactId>foundation-test-scaffolding</artifactId>
+    </dependency>
+  </dependencies>
 
 </project>
